Update solution.txt
Group 2 - Richard Remley
Comprehension 2
Completion 2

In theory this isn't difficult, but I found it absolutely impossible to move a repository or file up the command lines. I copied the templates as clearly as possible from online, and it did not ever seem to work, instead giving "No such file or directory" errors. Both being hyper specific and unspecific with the pathing was of no help.